Centralize element message notification in model base class
Move element notification logic from individual model implementations to a
common base class method to reduce code duplication and improve maintainability.

Changes:
- Moved meshXElementIF definition from meshx_element_class.hpp to
  meshx_fwd_decl.hpp for better forward declaration support
- Added prepare_element_msg() pure virtual method to meshXModelIF
- Added send_to_parent_element() common implementation to handle element
  state change and notification in one place
- Added update_element_state_change_header() virtual method for type-specific
  header casting in client/server models
- Refactored all model implementations (Config, Generic, Light) to store
  element messages as member variables and implement prepare_element_msg()
- Removed direct on_model_cb() calls from individual model implementations

This eliminates repetitive boilerplate code across 20+ model classes and
establishes a consistent pattern for element-state-to-model communication.

+++ main/component/meshx/elements/src/variants/meshx_relay_element.cpp
@@ -55,7 +55,7 @@ uint8_t meshXRelayServerElement MESHX_RELAY_SERVER_ELEMENT_TEMPLATE_PARAMS
     auto relay_model = std::make_unique<meshXGenericOnOffServerModel>(this, &element_ctx.gen_on_off_state);
     this->get_sig_models().push_back(std::move(relay_model));
 
-    return this->get_sig_models().size();
+    return (uint8_t)this->get_sig_models().size();
 }
 
 /**
@@ -113,7 +113,7 @@ uint8_t meshXRelayClientElement MESHX_RELAY_CLIENT_ELEMENT_TEMPLATE_PARAMS
     // Create Relay Client model
     auto relay_model = std::make_unique<meshXGenericOnOffClientModel>(this, &element_ctx.gen_on_off_state);
     this->get_sig_models().push_back(std::move(relay_model));
-    return this->get_sig_models().size();
+    return (uint8_t)this->get_sig_models().size();
 }
 
 /**
